Move site dataloader to graphql layer
