+++ saleor/graphql/order/mutations/draft_order_complete.py
@@ -15,12 +15,12 @@ from ....order.error_codes import OrderErrorCode
 from ....order.fetch import OrderInfo, OrderLineInfo
 from ....order.search import prepare_order_search_vector_value
 from ....order.utils import get_order_country
-from ....site.models import load_site
 from ....warehouse.management import allocate_preorders, allocate_stocks
 from ....warehouse.reservations import is_reservation_enabled
 from ...app.dataloaders import load_app
 from ...core.mutations import BaseMutation
 from ...core.types import OrderError
+from ...site.dataloaders import load_site
 from ..types import Order
 from ..utils import (
     prepare_insufficient_stock_order_validation_errors,
