+++ saleor/graphql/order/mutations/order_capture.py
@@ -7,11 +7,11 @@ from ....order.error_codes import OrderErrorCode
 from ....order.fetch import fetch_order_info
 from ....payment import PaymentError, TransactionKind, gateway
 from ....payment.gateway import request_charge_action
-from ....site.models import load_site
 from ...app.dataloaders import load_app
 from ...core.mutations import BaseMutation
 from ...core.scalars import PositiveDecimal
 from ...core.types import OrderError
+from ...site.dataloaders import load_site
 from ..types import Order
 from .utils import clean_payment, try_payment_action
